+++ changelog.md
@@ -14,6 +14,8 @@ Codion Change Log
 - SerializationWhitelist.DryRun, now excludes arrays and includes the component type.
 ### is.codion.swing.common.ui.tools
 - LoadTestPanel.loadTestPanel() factory method added, constructor now private.
+### is.codion.framework.domain
+- IdentityKeyGenerator now a singleton.
 ### is.codion.swing.framework.model.tools
 - DomainToString bug fixed, now handles the mapping from an unknown column data type to an object based Column correctly, DatabaseDomain some renaming.
 - DomainGeneratorModel, table type column moved.